Issue #1927650. Improve data for transaction.
Currently we let people set the transaction currency per payment instance, but we could instead get the currency from the order total which has that set.

Not making that change here - just getting the data available so we can use it in future, and modifying the way we get the purchase amount with decimals (or without).